feat: packaged-factories-directory-003 - Publish the canonical directory in customer guidance
